test(core): split watch sourcemap basic tests
