+++ packages/core/test/core/watch/source-maps-basic/helpers.ts
@@ -1,0 +1,13 @@
+export {
+  describe,
+  test,
+  expect,
+  watch,
+  mkdtempSync,
+  writeFileSync,
+  readFileSync,
+  rmSync,
+  existsSync,
+  join,
+  tmpdir,
+} from '../helpers';
